backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
pydantic-settings==2.7.0
yt-dlp==2024.12.23
openai-whisper==20240930
openai==1.58.1
python-multipart==0.0.20
